+++ src/callbacks.cc
@@ -29,6 +29,7 @@ uv_async_t async;
 //
 Nan::Callback *emit_cb;
 Nan::CopyablePersistentTraits<v8::Object>::CopyablePersistent ctx_obj;
+Nan::AsyncResource *resource;
 
 // Message passing queue between OpenZWave callback and v8 async handler.
 mutex zqueue_mutex;
@@ -137,7 +138,7 @@ void handleControllerCommand(NotifInfo *notif)
   info[2] = Nan::New<Integer>(notif->event);        // Driver::ControllerCommand
   info[3] = Nan::New<Integer>(notif->notification); // Driver::ControllerCommand
   info[4] = Nan::New<String>(notif->help.c_str()).ToLocalChecked();
-  emit_cb->Call(Nan::New(ctx_obj), 5, info);
+  emit_cb->Call(Nan::New(ctx_obj),  5, info, resource);
 }
 // ##### END OF LEGACY MODE ###### //
 #endif
@@ -171,7 +172,7 @@ void handleNotification(NotifInfo *notif)
     emitinfo[1] = Nan::New<Integer>(notif->nodeid);
     emitinfo[2] = Nan::New<Integer>(value.GetCommandClassId());
     emitinfo[3] = valobj;
-    emit_cb->Call(Nan::New(ctx_obj), 4, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  4, emitinfo, resource);
     break;
   }
   //                            ##################
@@ -192,7 +193,7 @@ void handleNotification(NotifInfo *notif)
     emitinfo[2] = Nan::New<Integer>(value.GetCommandClassId());
     emitinfo[3] = Nan::New<Integer>(value.GetInstance());
     emitinfo[4] = Nan::New<Integer>(value.GetIndex());
-    emit_cb->Call(Nan::New(ctx_obj), 5, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  5, emitinfo, resource);
     break;
   }
   //                            ##################
@@ -204,7 +205,7 @@ void handleNotification(NotifInfo *notif)
     emitinfo[1] = Nan::New<Integer>(notif->nodeid);
     emitinfo[2] = Nan::New<Integer>(value.GetCommandClassId());
     emitinfo[3] = valobj;
-    emit_cb->Call(Nan::New(ctx_obj), 4, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  4, emitinfo, resource);
     break;
   }
   //                            ####################
@@ -216,7 +217,7 @@ void handleNotification(NotifInfo *notif)
     emitinfo[1] = Nan::New<Integer>(notif->nodeid);
     emitinfo[2] = Nan::New<Integer>(value.GetCommandClassId());
     emitinfo[3] = valobj;
-    emit_cb->Call(Nan::New(ctx_obj), 4, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  4, emitinfo, resource);
     break;
   }
   //                            #############
@@ -242,7 +243,7 @@ void handleNotification(NotifInfo *notif)
     }
     emitinfo[0] = Nan::New<String>("node added").ToLocalChecked();
     emitinfo[1] = Nan::New<Integer>(notif->nodeid);
-    emit_cb->Call(Nan::New(ctx_obj), 2, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  2, emitinfo, resource);
     break;
   }
   //                            #################
@@ -254,7 +255,7 @@ void handleNotification(NotifInfo *notif)
     }
     emitinfo[0] = Nan::New<String>("node removed").ToLocalChecked();
     emitinfo[1] = Nan::New<Integer>(notif->nodeid);
-    emit_cb->Call(Nan::New(ctx_obj), 2, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  2, emitinfo, resource);
     break;
   }
   //                            ######################
@@ -273,7 +274,7 @@ void handleNotification(NotifInfo *notif)
     emitinfo[0] = Nan::New<String>("node naming").ToLocalChecked();
     emitinfo[1] = Nan::New<Integer>(notif->nodeid);
     emitinfo[2] = cbinfo;
-    emit_cb->Call(Nan::New(ctx_obj), 3, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  3, emitinfo, resource);
     break;
   }
   //                            ###############
@@ -283,7 +284,7 @@ void handleNotification(NotifInfo *notif)
     emitinfo[1] = Nan::New<Integer>(notif->nodeid);
     emitinfo[2] = Nan::New<Integer>(notif->event);
     emitinfo[3] = Nan::New<String>(notif->help.c_str()).ToLocalChecked();
-    emit_cb->Call(Nan::New(ctx_obj), 4, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  4, emitinfo, resource);
     break;
   }
   //                            #####################
@@ -293,7 +294,7 @@ void handleNotification(NotifInfo *notif)
       node->polled = false;
       emitinfo[0] = Nan::New<String>("polling disabled").ToLocalChecked();
       emitinfo[1] = Nan::New<Integer>(notif->nodeid);
-      emit_cb->Call(Nan::New(ctx_obj), 2, emitinfo);
+      emit_cb->Call(Nan::New(ctx_obj),  2, emitinfo, resource);
     }
     break;
   }
@@ -304,7 +305,7 @@ void handleNotification(NotifInfo *notif)
       node->polled = true;
       emitinfo[0] = Nan::New<String>("polling enabled").ToLocalChecked();
       emitinfo[1] = Nan::New<Integer>(notif->nodeid);
-      emit_cb->Call(Nan::New(ctx_obj), 2, emitinfo);
+      emit_cb->Call(Nan::New(ctx_obj),  2, emitinfo, resource);
     }
     break;
   }
@@ -314,7 +315,7 @@ void handleNotification(NotifInfo *notif)
     emitinfo[0] = Nan::New<String>("scene event").ToLocalChecked();
     emitinfo[1] = Nan::New<Integer>(notif->nodeid);
     emitinfo[2] = Nan::New<Integer>(notif->sceneid);
-    emit_cb->Call(Nan::New(ctx_obj), 3, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  3, emitinfo, resource);
     break;
   }
   //                            ##################
@@ -323,7 +324,7 @@ void handleNotification(NotifInfo *notif)
     emitinfo[0] = Nan::New<String>("create button").ToLocalChecked();
     emitinfo[1] = Nan::New<Integer>(notif->nodeid);
     emitinfo[2] = Nan::New<Integer>(notif->buttonid);
-    emit_cb->Call(Nan::New(ctx_obj), 3, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  3, emitinfo, resource);
     break;
   }
   //                            ##################
@@ -332,7 +333,7 @@ void handleNotification(NotifInfo *notif)
     emitinfo[0] = Nan::New<String>("delete button").ToLocalChecked();
     emitinfo[1] = Nan::New<Integer>(notif->nodeid);
     emitinfo[2] = Nan::New<Integer>(notif->buttonid);
-    emit_cb->Call(Nan::New(ctx_obj), 3, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  3, emitinfo, resource);
     break;
   }
   //                            ##############
@@ -341,7 +342,7 @@ void handleNotification(NotifInfo *notif)
     emitinfo[0] = Nan::New<String>("button on").ToLocalChecked();
     emitinfo[1] = Nan::New<Integer>(notif->nodeid);
     emitinfo[2] = Nan::New<Integer>(notif->buttonid);
-    emit_cb->Call(Nan::New(ctx_obj), 3, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  3, emitinfo, resource);
     break;
   }
   //                            ###############
@@ -350,7 +351,7 @@ void handleNotification(NotifInfo *notif)
     emitinfo[0] = Nan::New<String>("button off").ToLocalChecked();
     emitinfo[1] = Nan::New<Integer>(notif->nodeid);
     emitinfo[2] = Nan::New<Integer>(notif->buttonid);
-    emit_cb->Call(Nan::New(ctx_obj), 3, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  3, emitinfo, resource);
     break;
   }
   //                            #################
@@ -360,14 +361,14 @@ void handleNotification(NotifInfo *notif)
     homeid = notif->homeid;
     emitinfo[0] = Nan::New<String>("driver ready").ToLocalChecked();
     emitinfo[1] = Nan::New<Integer>(homeid);
-    emit_cb->Call(Nan::New(ctx_obj), 2, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  2, emitinfo, resource);
     break;
   }
   //                            ##################
   case OpenZWave::Notification::Type_DriverFailed: {
     //                            ##################
     emitinfo[0] = Nan::New<String>("driver failed").ToLocalChecked();
-    emit_cb->Call(Nan::New(ctx_obj), 1, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  1, emitinfo, resource);
     break;
   }
   //                            ##################################
@@ -377,7 +378,7 @@ void handleNotification(NotifInfo *notif)
     emitinfo[0] = Nan::New<String>("node available").ToLocalChecked();
     emitinfo[1] = Nan::New<Integer>(notif->nodeid);
     emitinfo[2] = cbinfo;
-    emit_cb->Call(Nan::New(ctx_obj), 3, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  3, emitinfo, resource);
     break;
   }
   /*
@@ -390,7 +391,7 @@ void handleNotification(NotifInfo *notif)
     emitinfo[0] = Nan::New<String>("node ready").ToLocalChecked();
     emitinfo[1] = Nan::New<Integer>(notif->nodeid);
     emitinfo[2] = cbinfo;
-    emit_cb->Call(Nan::New(ctx_obj), 3, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  3, emitinfo, resource);
     break;
   }
   /*
@@ -403,7 +404,7 @@ void handleNotification(NotifInfo *notif)
   case OpenZWave::Notification::Type_AllNodesQueriedSomeDead: {
     //                            #############################
     emitinfo[0] = Nan::New<String>("scan complete").ToLocalChecked();
-    emit_cb->Call(Nan::New(ctx_obj), 1, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  1, emitinfo, resource);
     break;
   }
   //                            ##################
@@ -413,7 +414,7 @@ void handleNotification(NotifInfo *notif)
     emitinfo[1] = Nan::New<Integer>(notif->nodeid);
     emitinfo[2] = Nan::New<Integer>(notif->notification);
     emitinfo[3] = Nan::New<String>(notif->help.c_str()).ToLocalChecked();
-    emit_cb->Call(Nan::New(ctx_obj), 4, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  4, emitinfo, resource);
     break;
   }
   case OpenZWave::Notification::Type_DriverRemoved:
@@ -432,7 +433,7 @@ void handleNotification(NotifInfo *notif)
     emitinfo[3] =
         Nan::New<Integer>(notif->notification); // Driver::ControllerState
     emitinfo[4] = Nan::New<String>(notif->help.c_str()).ToLocalChecked();
-    emit_cb->Call(Nan::New(ctx_obj), 5, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  5, emitinfo, resource);
     break;
   case OpenZWave::Notification::Type_NodeReset:
     emitinfo[0] = Nan::New<String>("node reset").ToLocalChecked();
@@ -440,7 +441,7 @@ void handleNotification(NotifInfo *notif)
     emitinfo[2] = Nan::New<Integer>(notif->event); // Driver::ControllerCommand
     emitinfo[3] =
         Nan::New<Integer>(notif->notification); // Driver::ControllerState
-    emit_cb->Call(Nan::New(ctx_obj), 4, emitinfo);
+    emit_cb->Call(Nan::New(ctx_obj),  4, emitinfo, resource);
     break;
 #endif
   default: