+++ src/openzwave-nodes.cc
@@ -31,7 +31,7 @@ namespace OZW {
 		Nan::HandleScope scope;
 		uint8* neighbors;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		uint8 numNeighbors = OpenZWave::Manager::Get()->GetNodeNeighbors(homeid, nodeid, &neighbors);
 		Local<Array> o_neighbors = Nan::New<Array>(numNeighbors);
 		if (numNeighbors > 0) {
@@ -49,7 +49,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		OpenZWave::Manager::Get()->SetNodeOn(homeid, nodeid);
 	}
 
@@ -59,7 +59,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		OpenZWave::Manager::Get()->SetNodeOff(homeid, nodeid);
 	}
 
@@ -122,8 +122,8 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(2, "nodeid, location");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
-		std::string location = (*String::Utf8Value(Nan::To<String>(info[1]).ToLocalChecked()));
+		uint8 nodeid = info[0]->Uint32Value();
+		std::string location(*Nan::Utf8String( info[1] ));
 		OpenZWave::Manager::Get()->SetNodeLocation(homeid, nodeid, location);
 	}
 
@@ -136,8 +136,8 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(2, "nodeid, name");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
-		std::string name = (*String::Utf8Value(Nan::To<String>(info[1]).ToLocalChecked()));
+		uint8 nodeid = info[0]->Uint32Value();
+		std::string name(*Nan::Utf8String( info[1] ));
 		OpenZWave::Manager::Get()->SetNodeName(homeid, nodeid, name);
 	}
 
@@ -157,7 +157,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		OpenZWave::Manager::Get()->RefreshNodeInfo(homeid, nodeid);
 	}
 
@@ -170,7 +170,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		std::string result = OpenZWave::Manager::Get()->GetNodeManufacturerName(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<String>(result.c_str()).ToLocalChecked());
 	}
@@ -184,7 +184,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		bool result = OpenZWave::Manager::Get()->RequestNodeState(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<Boolean>(result));
 	}
@@ -198,7 +198,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		bool result = OpenZWave::Manager::Get()->RequestNodeDynamic(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<Boolean>(result));
 	}
@@ -212,7 +212,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		bool result = OpenZWave::Manager::Get()->IsNodeListeningDevice(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<Boolean>(result));
 	}
@@ -226,7 +226,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		bool result = OpenZWave::Manager::Get()->IsNodeFrequentListeningDevice(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<Boolean>(result));
 	}
@@ -240,7 +240,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		bool result = OpenZWave::Manager::Get()->IsNodeBeamingDevice(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<Boolean>(result));
 	}
@@ -254,7 +254,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		bool result = OpenZWave::Manager::Get()->IsNodeRoutingDevice(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<Boolean>(result));
 	}
@@ -268,7 +268,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		bool result = OpenZWave::Manager::Get()->IsNodeSecurityDevice(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<Boolean>(result));
 	}
@@ -282,7 +282,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8  nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8  nodeid = info[0]->Uint32Value();
 		uint32 result = OpenZWave::Manager::Get()->GetNodeMaxBaudRate(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<Uint32>(result));
 	}
@@ -296,7 +296,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		uint8 result = OpenZWave::Manager::Get()->GetNodeVersion(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<Integer>(result));
 	}
@@ -310,7 +310,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		uint8 result = OpenZWave::Manager::Get()->GetNodeSecurity(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<Integer>(result));
 	}
@@ -324,7 +324,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		uint8 result = OpenZWave::Manager::Get()->GetNodeBasic(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<Integer>(result));
 	}
@@ -338,7 +338,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		uint8 result = OpenZWave::Manager::Get()->GetNodeGeneric(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<Integer>(result));
 	}
@@ -352,7 +352,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		uint8 result = OpenZWave::Manager::Get()->GetNodeSpecific(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<Integer>(result));
 	}
@@ -366,7 +366,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		std::string result = OpenZWave::Manager::Get()->GetNodeType(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<String>(result.c_str()).ToLocalChecked());
 	}
@@ -380,7 +380,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		std::string result = OpenZWave::Manager::Get()->GetNodeProductName(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<String>(result.c_str()).ToLocalChecked());
 	}
@@ -394,7 +394,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		std::string result = OpenZWave::Manager::Get()->GetNodeName(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<String>(result.c_str()).ToLocalChecked());
 	}
@@ -408,7 +408,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		std::string result = OpenZWave::Manager::Get()->GetNodeLocation(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<String>(result.c_str()).ToLocalChecked());
 	}
@@ -422,7 +422,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		std::string result = OpenZWave::Manager::Get()->GetNodeManufacturerId(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<String>(result.c_str()).ToLocalChecked());
 	}
@@ -436,7 +436,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		std::string result = OpenZWave::Manager::Get()->GetNodeProductType(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<String>(result.c_str()).ToLocalChecked());
 	}
@@ -450,7 +450,7 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
+		uint8 nodeid = info[0]->Uint32Value();
 		std::string result = OpenZWave::Manager::Get()->GetNodeProductId(homeid, nodeid);
 		info.GetReturnValue().Set(Nan::New<String>(result.c_str()).ToLocalChecked());
 	}
@@ -464,8 +464,8 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
-		std::string name = (*String::Utf8Value(Nan::To<String>(info[1]).ToLocalChecked()));
+		uint8 nodeid = info[0]->Uint32Value();
+		std::string name(*Nan::Utf8String( info[1] ));
 		OpenZWave::Manager::Get()->SetNodeManufacturerName(homeid, nodeid, name);
 	}
 
@@ -478,8 +478,8 @@ namespace OZW {
 	{
 		Nan::HandleScope scope;
 		CheckMinArgs(1, "nodeid");
-		uint8 nodeid = Nan::To<Number>(info[0]).ToLocalChecked()->Value();
-		std::string name = (*String::Utf8Value(Nan::To<String>(info[1]).ToLocalChecked()));
+		uint8 nodeid = info[0]->Uint32Value();
+		std::string name(*Nan::Utf8String( info[1] ));
 		OpenZWave::Manager::Get()->SetNodeProductName(homeid, nodeid, name);
 	}